Revised Login Form
Essentially just added an option for a user to logout and once logged in they can also see the rest of their information.

I plan on integrating this into the DOM in the near future.

Next I would like to figure out how to keep a user logged in after the page changes.